Handle auth for filters with references
