+++ adapters/handlers/grpc/v1/batch_delete.go
@@ -23,11 +23,14 @@ import (
 	"github.com/weaviate/weaviate/usecases/objects"
 )
 
-func batchDeleteParamsFromProto(req *pb.BatchDeleteRequest, getClass func(string) *models.Class) (objects.BatchDeleteParams, error) {
+func batchDeleteParamsFromProto(req *pb.BatchDeleteRequest, getClass func(string) (*models.Class, error)) (objects.BatchDeleteParams, error) {
 	params := objects.BatchDeleteParams{}
 
 	// make sure collection exists
-	class := getClass(req.Collection)
+	class, err := getClass(req.Collection)
+	if err != nil {
+		return params, err
+	}
 	if class == nil {
 		return objects.BatchDeleteParams{}, fmt.Errorf("could not find class %s in schema", req.Collection)
 	}
