+++ usecases/objects/batch_delete.go
@@ -35,7 +35,7 @@ func (b *BatchManager) DeleteObjects(ctx context.Context, principal *models.Prin
 	if match != nil {
 		class = match.Class
 	}
-	err := b.authorizer.Authorize(principal, authorization.UPDATE, authorization.Shards(class, tenant)...)
+	err := b.authorizer.Authorize(principal, authorization.DELETE, authorization.Shards(class, tenant)...)
 	if err != nil {
 		return nil, err
 	}
@@ -56,16 +56,11 @@ func (b *BatchManager) DeleteObjects(ctx context.Context, principal *models.Prin
 	return b.deleteObjects(ctx, principal, match, dryRun, output, repl, tenant)
 }
 
-// DeleteObjectsFromGRPC deletes objects in batch based on the match filter
-func (b *BatchManager) DeleteObjectsFromGRPC(ctx context.Context, principal *models.Principal,
+// DeleteObjectsFromGRPCAfterAuth deletes objects in batch based on the match filter
+func (b *BatchManager) DeleteObjectsFromGRPCAfterAuth(ctx context.Context, principal *models.Principal,
 	params BatchDeleteParams,
 	repl *additional.ReplicationProperties, tenant string,
 ) (BatchDeleteResult, error) {
-	err := b.authorizer.Authorize(principal, authorization.UPDATE, authorization.Shards(params.ClassName.String(), tenant)...)
-	if err != nil {
-		return BatchDeleteResult{}, err
-	}
-
 	unlock, err := b.locks.LockConnector()
 	if err != nil {
 		return BatchDeleteResult{}, NewErrInternal("could not acquire lock: %v", err)
@@ -143,7 +138,7 @@ func (b *BatchManager) validateBatchDelete(ctx context.Context, principal *model
 		return nil, 0, fmt.Errorf("failed to parse where filter: %s", err)
 	}
 
-	err = filters.ValidateFilters(b.schemaManager.ReadOnlyClass, filter)
+	err = filters.ValidateFilters(b.classGetterFunc(principal), filter)
 	if err != nil {
 		return nil, 0, fmt.Errorf("invalid where filter: %s", err)
 	}
@@ -165,4 +160,17 @@ func (b *BatchManager) validateBatchDelete(ctx context.Context, principal *model
 		Output:    outputParam,
 	}
 	return params, vclasses[match.Class].Version, nil
+}
+
+func (b *BatchManager) classGetterFunc(principal *models.Principal) func(string) (*models.Class, error) {
+	return func(name string) (*models.Class, error) {
+		if err := b.authorizer.Authorize(principal, authorization.READ, authorization.Collections(name)...); err != nil {
+			return nil, err
+		}
+		class := b.schemaManager.ReadOnlyClass(name)
+		if class == nil {
+			return nil, fmt.Errorf("could not find class %s in schema", name)
+		}
+		return class, nil
+	}
 }
