+++ usecases/traverser/traverser_aggregate.go
@@ -18,7 +18,6 @@ import (
 	"github.com/pkg/errors"
 	"github.com/weaviate/weaviate/entities/aggregation"
 	enterrors "github.com/weaviate/weaviate/entities/errors"
-	"github.com/weaviate/weaviate/entities/filters"
 	"github.com/weaviate/weaviate/entities/models"
 	"github.com/weaviate/weaviate/usecases/auth/authorization"
 	"github.com/weaviate/weaviate/usecases/modules"
@@ -43,6 +42,11 @@ func (t *Traverser) Aggregate(ctx context.Context, principal *models.Principal,
 	defer unlock()
 
 	inspector := newTypeInspector(t.schemaGetter.ReadOnlyClass)
+
+	// validate here, because filters can contain references that need to be authorized
+	if err := t.validateFilters(principal, params.Filters); err != nil {
+		return nil, errors.Wrap(err, "invalid 'where' filter")
+	}
 
 	if params.NearVector != nil || params.NearObject != nil || len(params.ModuleParams) > 0 {
 		className := params.ClassName.String()
@@ -106,11 +110,6 @@ func (t *Traverser) Aggregate(ctx context.Context, principal *models.Principal,
 		params.Certainty = certainty
 	}
 
-	if params.Filters != nil {
-		if err := filters.ValidateFilters(t.schemaGetter.ReadOnlyClass, params.Filters); err != nil {
-			return nil, errors.Wrap(err, "invalid 'where' filter")
-		}
-	}
 	var mp *modules.Provider
 
 	if t.nearParamsVector.modulesProvider != nil {
