+++ usecases/traverser/traverser_get.go
@@ -16,6 +16,9 @@ import (
 	"fmt"
 	"time"
 
+	"github.com/pkg/errors"
+	"github.com/weaviate/weaviate/entities/filters"
+
 	"github.com/weaviate/weaviate/entities/dto"
 	enterrors "github.com/weaviate/weaviate/entities/errors"
 	"github.com/weaviate/weaviate/entities/models"
@@ -56,6 +59,11 @@ func (t *Traverser) GetClass(ctx context.Context, principal *models.Principal,
 		return nil, enterrors.NewErrLockConnector(err)
 	}
 	defer unlock()
+
+	// validate here, because filters can contain references that need to be authorized
+	if err := t.validateFilters(principal, params.Filters); err != nil {
+		return nil, errors.Wrap(err, "invalid 'where' filter")
+	}
 
 	certainty := ExtractCertaintyFromParams(params)
 	if certainty != 0 || params.AdditionalProperties.Certainty {
@@ -99,4 +107,21 @@ func (t *Traverser) probeForRefDepthLimit(props search.SelectProperties) error {
 		return fmt.Errorf("nested references depth exceeds QUERY_CROSS_REFERENCE_DEPTH_LIMIT (%d)", depthLimit)
 	}
 	return nil
+}
+
+func (t *Traverser) validateFilters(principal *models.Principal, filter *filters.LocalFilter) error {
+	if filter == nil {
+		return nil
+	}
+
+	f := func(name string) (*models.Class, error) {
+		err := t.authorizer.Authorize(principal, authorization.READ, authorization.Collections(name)...)
+		if err != nil {
+			return nil, err
+		}
+
+		return t.schemaGetter.ReadOnlyClass(name), nil
+	}
+
+	return filters.ValidateFilters(f, filter)
 }
